Added slow-path credentials logger
